cleanup FE
