+++ client/package.json
@@ -30,7 +30,7 @@
     "@vitejs/plugin-react": "^4.3.4",
     "@vitest/coverage-v8": "^3.0.9",
     "autoprefixer": "^10.4.21",
-    "eslint": "^9.21.0",
+    "eslint": "^9.36.0",
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^15.15.0",
