+++ client/src/components/ForecastDisplay.jsx
@@ -1,4 +1,9 @@
 import PropTypes from "prop-types";
+import {
+  getConditionName,
+  groupForecastsByDay,
+  formatDate,
+} from "../utils/forecastUtils";
 
 const ForecastDisplay = ({ data }) => {
   if (!data || !data.forecastTimestamps) return null;
@@ -59,59 +64,6 @@ const ForecastDisplay = ({ data }) => {
       </div>
     </div>
   );
-};
-
-const formatDate = (dateString) => {
-  const date = new Date(dateString);
-  return date.toLocaleDateString("en-US", {
-    weekday: "short",
-    month: "short",
-    day: "numeric",
-  });
-};
-
-// move-BE
-const groupForecastsByDay = (forecasts) => {
-  const days = {};
-
-  forecasts.forEach((forecast) => {
-    const date = new Date(forecast.forecastTimeUtc);
-    const dayKey = date.toISOString().split("T")[0];
-    const hour = date.getHours();
-
-    if (!days[dayKey] || (hour >= 12 && hour <= 14)) {
-      days[dayKey] = forecast;
-    }
-  });
-
-  return days;
-};
-
-// move-BE
-const getConditionName = (code) => {
-  const conditions = {
-    clear: "Clear",
-    "partly-cloudy": "Partly Cloudy",
-    "cloudy-with-sunny-intervals": "Cloudy with Sunny Intervals",
-    cloudy: "Cloudy",
-    "light-rain": "Light Rain",
-    rain: "Rain",
-    "heavy-rain": "Heavy Rain",
-    thunder: "Thunder",
-    "isolated-thunderstorms": "Isolated Thunderstorms",
-    thunderstorms: "Thunderstorms",
-    "heavy-rain-with-thunderstorms": "Heavy Rain with Thunderstorms",
-    "light-sleet": "Light Sleet",
-    sleet: "Sleet",
-    "freezing-rain": "Freezing Rain",
-    hail: "Hail",
-    "light-snow": "Light Snow",
-    snow: "Snow",
-    "heavy-snow": "Heavy Snow",
-    fog: "Fog",
-  };
-
-  return conditions[code] || "Unknown";
 };
 
 ForecastDisplay.propTypes = {
